FIX Constants empty spaces and move to utils
