+++ app/src/main/java/com/codingpit/pvpcplanner/utils/Constants.kt
@@ -1,4 +1,4 @@
-package com.codingpit.pvpcplanner.data
+package com.codingpit.pvpcplanner.utils
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
@@ -11,13 +11,9 @@ object Constants {
     const val NOTIFICATION_SCREEN = "Notification"
     const val SETTING_SCREEN = "Setting"
 
-
     val ROUTES_NAVIGATION_BOTTOM_BAR = mapOf(
         HOME_SCREEN to Icons.Default.Home,
         NOTIFICATION_SCREEN to Icons.Default.Notifications,
         SETTING_SCREEN to Icons.Default.Settings,
     )
-
-
-
 }
